+++ src/tests/test_magic.py
@@ -2345,3 +2345,87 @@ def test_json_arrow_operators_with_snippets(ip, query_save, query_snippet, expec
     result = ip.run_cell(query_snippet).result
     result = list(result.dict().values())[0][0]
     assert result == expected
+
+
+@pytest.mark.parametrize(
+    "query, expected",
+    [
+        (
+            """%%sql
+SELECT 1""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1 -- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1
+-- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment""",
+            1,
+        ),
+        (
+            """%%sql
+-- comment before
+SELECT 1;""",
+            1,
+        ),
+        (
+            """%%sql
+-- comment before
+SELECT 1;
+-- comment after""",
+            1,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment
+SELECT 2""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment
+SELECT 2;""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment
+SELECT 2;""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1;
+-- comment before
+SELECT 2;
+-- comment after""",
+            2,
+        ),
+        (
+            """%%sql
+SELECT 1; -- comment before
+SELECT 2;
+-- comment after""",
+            2,
+        ),
+    ],
+)
+def test_query_comment_after_semicolon(ip, query, expected):
+    result = ip.run_cell(query).result
+    assert list(result.dict().values())[-1][0] == expected